public & private API
