--- base/rico-core/src/main/java/dev/rico/core/http/HttpExecutorFactory.java
@@ -1,7 +1,0 @@
-package dev.rico.core.http;
-
-import java.util.function.Function;
-
-@FunctionalInterface
-public interface HttpExecutorFactory<R> extends Function<HttpProvider<R>, HttpExecutor<R>> {
-}
