+++ base/rico-core/src/main/java/dev/rico/internal/core/http/AbstractHttpCallRequestBuilder.java
@@ -1,6 +1,10 @@
-package dev.rico.core.http;
+package dev.rico.internal.core.http;
 
 import com.google.gson.Gson;
+import dev.rico.core.http.ByteArrayProvider;
+import dev.rico.core.http.HttpCallRequestBuilder;
+import dev.rico.core.http.HttpCallResponseBuilder;
+import dev.rico.core.http.HttpURLConnectionHandler;
 import dev.rico.internal.core.Assert;
 import dev.rico.internal.core.http.HttpClientConnection;
 import dev.rico.internal.core.http.HttpHeaderImpl;
