+++ base/rico-core/src/main/java/dev/rico/internal/core/http/AbstractHttpClient.java
@@ -1,6 +1,7 @@
-package dev.rico.core.http;
+package dev.rico.internal.core.http;
 
 import com.google.gson.Gson;
+import dev.rico.core.http.*;
 import dev.rico.internal.core.Assert;
 import dev.rico.internal.core.http.DefaultHttpURLConnectionFactory;
 import dev.rico.internal.core.http.HttpClientConnection;
