+++ base/rico-core/src/main/java/dev/rico/internal/core/http/HttpProvider.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.rico.core.http;
+package dev.rico.internal.core.http;
 
 import dev.rico.core.http.HttpException;
 import dev.rico.core.http.HttpResponse;
